config (backend): config query

+++ src/main/java/com/project/t8/repository/LogRepo.java
@@ -12,9 +12,9 @@ import java.util.List;
 @Repository
 public interface LogRepo extends JpaRepository<Log,Long>{
     @Query(value = "select d from Log d where d.userId = :userId")
-    List<Log> findByUserId(@Param("userId") Long userId);
+    List<Log> findByUserId(@Param("userId") int userId);
     @Query("SELECT l FROM Log l WHERE (:userId IS NULL OR l.userId = :userId) and FUNCTION('MONTH', l.createdAt) = :month")
-    List<Log> findByMonth(@Param("userId") Long userId,@Param("month") int month);
+    List<Log> findByMonth(@Param("userId") int userId,@Param("month") int month);
     List<Log> findByDepartmentId(Long departmentId);
     Log findByLogId(Long logId);
 }
